meld: Fix build on Darwin
Meld’s post install script requires gtk-update-icon-cache program on PATH.
Because Meld is a Python app, it uses strictDeps so we need to explicitly
add the program to nativeBuildInputs.

For some reason, Linux version has the program available, though,
which hides the missing dependency. Perhaps it is propagated by something
but I do not know how to determine it.